requirements_dev.txt
-e .
pymongo[srv]

pylint
pre-commit
pytest
pytest-cov
black
pyhamcrest
flake8
python-dotenv

build
